将与 KVStore 相关的源文件移至专用的 kvstore 目录，在 Client 类中实现了延迟基准测试功能，添加了Config用以快速调整参数，修改CMakeLists

+++ src/kvstore/HashTable.cpp
@@ -67,7 +67,7 @@ void HashTable<T>::rehash() {
     bucketsz <<= 1;
     decltype(buckets) oldbuckets(bucketsz);
     swap(oldbuckets, buckets);
-    for (int i = 0; i < oldbucketsz; i++) {
+    for (size_t i = 0; i < oldbucketsz; i++) {
         auto &node = oldbuckets[i];
         if (node.status != OCCUPIED)
             continue;
